requirements.txt
# common
mysqlclient
django
django-redis

# drf
djangorestframework
drf-yasg
djangorestframework-jwt

# yuque
requests